requirements.txt
    mysql-connector-python